chore: rename lineups to V5 structure;
