+++ src/lib/components/page/fixture/lineups/Lineups-Vector-Mobile-Away.svelte
