+++ src/lib/components/page/fixture/lineups/Lineups-Vector-Tablet.svelte
